third_party/blink/web_tests/platform/generic/fast/dom/frameElement-accessor-context-expected.txt

Accessing window.frameElement from inside an iframe should not cause the <iframe>'s prototype to come from its own context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS iframe instanceof HTMLIFrameElement is true
PASS iframe instanceof iframe.contentWindow.HTMLIFrameElement is false
PASS successfullyParsed is true

TEST COMPLETE
